charts/pythweb/script/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.9
schedule==1.1.0
Telethon==1.25.1
Werkzeug==2.2.2
zipp==3.11.0
